Upgrade to the latest TypeScript version
I realized that the TypeScript version built into WebStorm was different
(and newer, as of the latest release) than the version that I have
globally installed, and both were out of date.

This is the core of some weird issues I've been seeing lately, where
I revert a git change made from files I hand compiled and then
WebStorm decides that they're slightly different.

To get around this, I've upgraded my global compiler to the latest
version and told WebStorm to use that version instead of its included
version.

This required some slight changes to the tsconfig.json files because
I'm generating a single output file but the config is specifying a
module format that does not allow this. That doesn't actually matter
since I'm not using modules anyway, but changing it makes the
compiler happy so there you have it.