Extend the Point class a bit
Add the ability to pass Vector instances to some methods, such as
translate, since that is totally a thing that I wish already existed
and is natural.

While in the vicinity I remembered that I made a utility function to
clamp values to a range, so I modified the implementation of
clampX/clampY to use that, for cleanliness.