Extend the Vector2D class a bit
As an oversight/oversimplification on my part, I made the vector API's
relating to its direction take degrees.

This is great from a debugging standpoint but awful from a using
vectors standpoint, since it requires a bunch of extra conversions
to/from radians, which is a drag and a half.

So now all of the API's that would take or return a direction default
naturally to radians and have a degrees version that does a
converstion for you.

Also, since there is a Utils function for doing these conversions,
they're used now instead of inlining it all the time.