Typos and global stages
Fixed a couple of typos while adding the notion that nurdz.stage
contains a reference to the last constructed Stage object, which
should hopefully be the *only* one ever created.

This aids in debugging because it allows console access to the stage,
and through that the scenes and their contents.