Fix a bug in vector magnitude code
The setter for the Vector2d.magnitude property uses trig to adjust
the magnitude of the vector while retaining its direction.

There was a "flaw" in the code in which if a negative magnitude was
provided, the operation would work as expected to set the magnitude
but would also reverse the direction of the vector.

This is fixed by ensuring that we only use the absolute value of the
magnitude value provided.