Bump the engine to version 0.3.0
This will be the version of the engine that will be used for the next
project in the series from the Chris DeLeon Udemy game development
course, "brick breaker" i.e. breakout.

This includes one last bug fix from the 0.2 series used for ts-tennis
that uncovered a bug in the vector magnitude code regarding allowing
a negative magnitude to be set, which would flip the vector direction
unintentionally.