Include support for touch events and scaling
Here we add in two new features; scaling the canvas up to fill the
page as much as possible, and touch events.

Currently, touch events are always handled, although unless the device
being used can generate them, they have no effect. The touch events
are converted into appropriate mouse events and triggered that way,
so anything done with a mouse should also be possible via a touch.

The canvas scaling is actually a two part thing; firstly, the canvas
is now properly centered horizontally AND vertically inside the page,
taking into account the size of the headers and footers (also the
footer is now properly static).

Additionally, an optional boolean (which defaults to false) can be
specified, which causes the canvas to visually scale to fill the whole
window container less the header and footer, while maintaining aspect
ratio.

When the boolean is false, the canvas still centers itself in the page
when the orientation of the device or the size of the window changes.